Implement the test of the feature Buy weighted item

+++ src/main/java/com/yd/kata/sp/model/Promotion.java
@@ -4,5 +4,5 @@ import java.math.BigDecimal;
 
 public interface Promotion {
 
-    public abstract BigDecimal computePriceWithPromotion(int itemQuantity, BigDecimal itemPrice);
+    public abstract BigDecimal computePriceWithPromotion(BigDecimal itemQuantity, BigDecimal itemPrice);
 }
